onnxruntime/requirements.txt
# Python Dependencies für ONNX Runtime App
onnxruntime>=1.23.1
flask>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
opencv-python>=4.8.0
rembg>=2.0.0